추가 double

+++ JavaStudy_jaewon/src/j22_람다식/Math.java
@@ -2,5 +2,5 @@ package j22_람다식;
 
 @FunctionalInterface
 public interface Math<T1, T2> {
-	public int calc(T1 value1, T2 value2);
+	public double calc(T1 value1, T2 value2);
 }
